support retry on browser for better performance
Currently, we use rod.Sleeper to abstract the retry so that we can use Go code to control the logic, it's very flexible but each retry will cost an RPC request to the browser.

We should add a retry helper to helper.js to let users do the retry on the browser.

We should design a way to cancel the remote wait to prevent leaking. Not like puppeteer's page.waitForFunction, it has no way to abort the wait before the timeout which can easily cause leaking on high load.
One possible design will be returning a promise and a cancel function together, like return { job, cacnel }.

However, the performance gain in real-world websites might be trivial. Because usually, only the first element will need retry.
After the first element is rendered, others will also be ready for query.

resolve #371